Converted GameEntityLoader to TypeScript.
